app/requirements.txt
django==2.2
djangorestframework==3.9.2
psycopg2==2.8
pytest-django==3.4.8
pytest-cov==2.6.1
requests-mock==1.5.2
requests==2.21.0
